feat: complete Seagull Storm - asset foundation, prefabs, full scenes, pixel UI
- deterministic .meta GUIDs for all project assets incl. sliced sprite
  sheets (Point filter, no compression, grid slicing per design spec)
- mascot pixel-art sprites ported from the Godot example; new real
  tilemap.png + ui.png (shared mascot-style art)
- all gameplay/UI prefabs authored (4 enemies, projectile, XP shell,
  poison zone, upgrade slot, leaderboard/news entries, levelup card)
- BootScene/TitleScene/GameScene rebuilt and fully wired (every
  SerializeField from the audit inventory), EditorBuildSettings fixed
- URP 2D pipeline bound with PixelPerfectCamera (480x270), Press Start
  2P TMP font asset, TMP essentials committed (shader dependency)
- manifest: standalone TMP removed (Unity 6 ugui), input system 1.19.0,
  broken module ref dropped; ProjectVersion pinned 6000.5.0f1
- GameBootstrap creates HorizonManager when missing (U7)
- Editor/SeagullSetup.ValidateSetup batch gate: 159 wiring assertions
  green, zero compile errors
- ConfigImporter API-key output path gitignored

+++ Assets/Scripts/UI/Screens/HubController.cs
@@ -135,7 +135,7 @@ namespace SeagullStorm
                 var obj = Instantiate(leaderboardEntryPrefab, leaderboardContainer);
                 var comp = obj.GetComponent<LeaderboardEntry>();
                 if (comp != null)
-                    comp.Setup(entry.position, entry.username, entry.score);
+                    comp.Setup((int)entry.position, entry.username, entry.score);
             }
         }
